graph: Fix silent truncation in BlockPtr::from(BlockWrapper)
Use From<(B256, u64)> which has proper i32::try_from().unwrap()
overflow checking instead of direct `as i32` cast that silently
truncates.

+++ graph/src/blockchain/types.rs
@@ -202,13 +202,13 @@ impl slog::Value for BlockPtr {
 
 impl From<BlockWrapper> for BlockPtr {
     fn from(b: BlockWrapper) -> BlockPtr {
-        BlockPtr::from((b.hash(), b.number_u64() as i32))
+        BlockPtr::from((b.hash(), b.number_u64()))
     }
 }
 
 impl From<&BlockWrapper> for BlockPtr {
     fn from(b: &BlockWrapper) -> BlockPtr {
-        BlockPtr::from((b.hash(), b.number_u64() as i32))
+        BlockPtr::from((b.hash(), b.number_u64()))
     }
 }
